changed type of count from double to float

+++ src/de/typology/trainers/LuceneTypologyIndexer.java
@@ -8,8 +8,8 @@ import java.util.ArrayList;
 import org.apache.lucene.analysis.Analyzer;
 import org.apache.lucene.analysis.core.SimpleAnalyzer;
 import org.apache.lucene.document.Document;
-import org.apache.lucene.document.DoubleField;
 import org.apache.lucene.document.Field;
+import org.apache.lucene.document.FloatField;
 import org.apache.lucene.document.StringField;
 import org.apache.lucene.index.IndexWriter;
 import org.apache.lucene.index.IndexWriterConfig;
@@ -46,7 +46,7 @@ public class LuceneTypologyIndexer {
 
 	private String line;
 	private String[] lineSplit;
-	private Double edgeCount;
+	private Float edgeCount;
 
 	public LuceneTypologyIndexer(String edgeDir) throws IOException {
 		Directory dir = FSDirectory.open(new File(edgeDir));
@@ -91,8 +91,8 @@ public class LuceneTypologyIndexer {
 				IOHelper.strongLog("lineSplit length is "
 						+ this.lineSplit.length);
 			}
-			this.edgeCount = Double
-					.parseDouble(this.lineSplit[this.lineSplit.length - 1]
+			this.edgeCount = Float
+					.parseFloat(this.lineSplit[this.lineSplit.length - 1]
 							.substring(1));
 			Document document = this.getDocument(this.lineSplit[0],
 					this.lineSplit[1], this.edgeCount);
@@ -102,11 +102,11 @@ public class LuceneTypologyIndexer {
 		return docCount;
 	}
 
-	private Document getDocument(String source, String target, Double count) {
+	private Document getDocument(String source, String target, Float count) {
 		Document document = new Document();
 		document.add(new StringField("src", source, Field.Store.YES));
 		document.add(new StringField("tgt", target, Field.Store.YES));
-		document.add(new DoubleField("cnt", count, Field.Store.YES));
+		document.add(new FloatField("cnt", count, Field.Store.YES));
 		return document;
 	}
 }
